backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
celery[redis]==5.4.0
redis==5.0.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.9.0
email-validator==2.1.0
pydantic-settings==2.5.0
PyPDF2==3.0.1
python-docx==1.1.2
openpyxl==3.1.5
slowapi==0.1.9
httpx==0.27.0
pytest==8.3.0
pytest-asyncio==0.24.0
python-dotenv==1.0.1
aiofiles==24.1.0
